starvation_monitor: Implement the conservative mode
Instead of dispatching one thread per CPU, monitors all CPUs from a
single thread. Once a starving thread is detected for a half of the
threshold time, start a thread to monitor that CPU.

If a thread monitoring a CPU detects that there is nothing to do for 10
seconds, it stops monitoring the CPU, returning to monitoring for the
main thread.

the -A/--aggressive_mode knob forces the creation of a monitoring thread
per CPU.

Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>