use FIFO for boosting (v3)
This is a set of changes that allow stalld to operation without the
Deadline scheduler policy. The idea is to emulate how deadline
works boosting a starving thread. When a boost operation is called
for, the fifo boosting logic figures out how many 'periods' would
be in the specified boost duration (default 3 seconds) and iterates
for that number of periods, boosting the starving thread to fifo:98,
sleeping for the 'runtime', deboosting to the original policy, sleeping
for the remainder (period - runtime) and repeating until the boost
duration is done.

The FIFO boost logic is turned on when the running kernel does not
support DEADLINE or when the command line option -F/--force_fifo is
specified.

Update error handling and behavior if unable to change policies so
that daemon runs in logging only mode.

Temporarily changed dl runtime during check_policies so that deadline
runtime is half of period. Restore after successfull boost.

Signed-off-by: Clark Williams <williams@redhat.com>