implement RT throttling management and refactor source files
This got a little bigger than I intended. Originally I wanted
to just add a function to disable RT throttling while running
and restore it on exit, but it grew to some refactoring when
I added the signal handler, then moved functions into separate
files.

The function turn_off_rt_throttling() just reads the file
/proc/sys/kernel/sched_rt_runtime_us and saves it's value. If
the value is -1, no action is performed, but if it's > 0 then
we save it and write a -1 to disable throttling, then register
an on_exit handler to restore it on exit.

That's when I realized that we don't shutdown gracefully on signals
so added a 'running' variable for loop termination and set that to
zero in the signal handler.

The signal handling logic and the parser utility functions moved
into src/utils.c. The throttling logic is in src/throttling.c.
I commented out the daemonize() signal calls to prevent conflict
with the other signal handling and then added a couple
of log_msg calls to indicate boost operations happening. Finally,
changed conservative_main() and aggressive_main() to be void returns
since we never check their return value.

There will be more refactoring to come since I want to do a little
namespace cleanup/protection using static (poor-man's C++).

Signed-off-by: Clark Williams <williams@redhat.com>